Refactor and optimize perfume normalizer module
In this commit, the perfume normalizer module has been refactored and optimized. Code related to the normalziation of French names was extracted into a new class named "FrenchNameNormalizer" to improve code organization and readability. Additionally, the program's capability to handle edge cases was significantly improved. The code for filling missing values was upgraded to better accommodate 'NaN' and 'None' cases. The sequence of code execution was also rearranged to enhance performance. This revised module is expected to be more efficient and robust than its predecessor.